requirements.txt
streamlit
crewai
crewai-tools
langchain-groq
python-dotenv
pydantic
